backend/requirements.txt
fastapi
uvicorn[standard]
imagehash
Pillow
faiss-cpu
numpy
easyocr
deepface
open-clip-torch
opencv-python
librosa
google-generativeai
geopy
faker
python-multipart
sqlalchemy
